# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Ortus Source Query API
  description: |
    REST API für die Abfrage von Datenquellen mittels Punktkoordinaten.

    Ortus ermöglicht die effiziente räumliche Abfrage von Datenquellen und unterstützt
    verschiedene Koordinatensysteme (SRIDs). Die API kann mehrere Datenquellen gleichzeitig
    verwalten und abfragen.

    ## Koordinatensysteme

    Die API unterstützt verschiedene Koordinatensysteme über den `srid` Parameter:
    - **4326** (WGS84): Standard GPS-Koordinaten (lon/lat)
    - **3857** (Web Mercator): Web-Mapping Standard
    - **25832/25833** (ETRS89/UTM): Deutscher/EU Standard
    - **31466/31467** (DHDN/Gauß-Krüger): Deutsches Legacy-System

    ## Authentifizierung

    Derzeit ist keine Authentifizierung erforderlich.
  version: 1.0.0
  contact:
    name: Ortus API Support
    url: https://github.com/jobrunner/ortus
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: API Version 1
  - url: /
    description: Root (für Health-Endpoints)
...
paths:
...
  /gazetteer:
    get:
      tags:
        - Gazetteer
      summary: Reverse-Geocoding + Peilung
      description: |
        Löst eine Koordinate auf ihre Verwaltungshierarchie (`admin`) auf und berechnet
        eine Peilung zum salientesten nahen Ort (`bearing`, z. B. "4 km E Würzburg").
        Jeder Teil ist `null`, wenn es kein Ergebnis gibt — keine Admin-Abdeckung bzw.
        kein Anker in Reichweite. Nur verfügbar, wenn das Gazetteer-Feature aktiv ist.
        Ein projizierter `srid` (z. B. 3857) wird intern nach WGS84 reprojiziert; die
        Antwort enthält die reprojizierte Koordinate im `wgs84`-Block. Nur ein `srid`,
        der sich nicht nach WGS84 transformieren lässt, wird mit 422 abgelehnt.
      operationId: gazetteer
      parameters:
        - $ref: '#/components/parameters/LonParam'
        - $ref: '#/components/parameters/LatParam'
        - $ref: '#/components/parameters/XParam'
        - $ref: '#/components/parameters/YParam'
        - $ref: '#/components/parameters/SridParam'
      responses:
        '200':
          description: Erfolgreiche Abfrage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GazetteerResponse'
              example:
                coordinate:
                  x: 9.93
                  y: 49.79
                  srid: 4326
                wgs84:
                  lon: 9.93
                  lat: 49.79
                admin:
                  country_iso: DE
                  hierarchy:
                    - level: 8
                      name: Würzburg
                      name_native: ""          # empty: already-Latin name (identical to name)
                      name_source: latin-osm
                      equivalent: municipality
                      local_term: Kreisfreie Stadt
                      equivalent_description: Local municipal authority (city/town/commune).
                    - level: 4
                      name: Bayern
                      name_native: ""          # empty: already-Latin name (identical to name)
                      name_source: latin-osm
                      equivalent: state
                      local_term: Land
                      equivalent_description: First-level subdivision (state, region, province).
                islands: null             # null: Würzburg is inland (an array of islands when the point is on one)
                bearing:
                  reference: Würzburg
                  name_native: ""            # empty: already-Latin name (identical to reference)
                  name_source: latin-osm
                  class: city
                  distance_km: 4.0
                  azimuth: 90.0
                  compass: E
                  label: "4 km E Würzburg"
                  inside: false
                exposure:                 # null when no DEM is wired or the point/neighbour lacks coverage
                  slope_deg: 12.5
                  slope_percent: 22.2
                  aspect_deg: 135.0       # null when flat
                  aspect_compass: SE      # "" when flat
                  flat: false
                  sample_spacing_m: 30.0
                  source:
                    name: "Copernicus DEM GLO-30"
                    url: "https://doi.org/10.5270/ESA-c5d3d65"
                    attribution: "© DLR e.V. 2010-2014 und © Airbus Defence and Space GmbH 2014-2018 provided under COPERNICUS by the European Union and ESA; all rights reserved."
                elevation:                # null when no DEM is configured
                  meters: 177.0
                  accuracy_m: 4.0
                  accuracy_basis: "GLO-30 LE90 (absolute)"
                  horizontal_accuracy_m: 6.0
                  vertical_datum: "EGM2008"
                  sea_level: false
                  surface_model: "DSM"
                  source:
                    name: "Copernicus DEM GLO-30"
                    url: "https://doi.org/10.5270/ESA-c5d3d65"
                    attribution: "© DLR e.V. 2010-2014 und © Airbus Defence and Space GmbH 2014-2018 provided under COPERNICUS by the European Union and ESA; all rights reserved."
                sources:
                  - code: latin-osm
                    short: OSM name (already Latin)
                    long: Taken verbatim from the OpenStreetMap name tag; already Latin script, no transliteration applied.
                    standard: ""    # latin-osm has no romanization standard (null in the manifest)
                license:
                  name: "ODbL-1.0"
                  url: "https://opendatacommons.org/licenses/odbl/1-0/"
                  attribution: "© OpenStreetMap contributors (ODbL 1.0); Natural Earth (public domain); GeoNames (CC BY 4.0); NGA GNS (public domain)"
        '400':
          description: Ungültige Parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Interner Serverfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sources:
    get:
      tags:
        - Sources
      summary: Alle Datenquellen auflisten
      description: Gibt eine Liste aller registrierten Datenquellen zurück.
      operationId: listSources
      responses:
        '200':
          description: Liste der Datenquellen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceList'
              example:
                sources:
                  - id: districts
                    name: districts.gpkg
                    path: /data/districts.gpkg
                    size: 1048576
                    layer_count: 2
                    indexed: true
                    ready: true
                    loaded_at: "2024-01-15T10:30:00Z"
                    last_queried: "2024-01-15T10:35:00Z"
                count: 1
        '500':
          description: Interner Serverfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sources/{sourceId}:
    get:
      tags:
        - Sources
      summary: Datenquellen-Details abrufen
      description: Gibt detaillierte Informationen zu einer bestimmten Datenquelle zurück.
      operationId: getSource
      parameters:
        - $ref: '#/components/parameters/SourceIdParam'
      responses:
        '200':
          description: Datenquellen-Details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          description: Datenquelle nicht gefunden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Interner Serverfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    SourceIdParam:
      name: sourceId
      in: path
      required: true
      description: Die eindeutige ID der Datenquelle
      schema:
        type: string
      example: districts
    LonParam:
      name: lon
      in: query
      description: Longitude (Längengrad) für WGS84-Koordinaten
      schema:
        type: number
        format: double
        minimum: -180
        maximum: 180
      example: 13.405
    LatParam:
      name: lat
      in: query
      description: Latitude (Breitengrad) für WGS84-Koordinaten
      schema:
        type: number
        format: double
        minimum: -90
        maximum: 90
      example: 52.52
    XParam:
      name: x
      in: query
      description: X-Koordinate für andere Koordinatensysteme
      schema:
        type: number
        format: double
      example: 389283
    YParam:
      name: y
      in: query
      description: Y-Koordinate für andere Koordinatensysteme
      schema:
        type: number
        format: double
      example: 5819450
    SridParam:
      name: srid
      in: query
      description: |
        Spatial Reference ID der Eingabekoordinaten.
        Standard ist 4326 (WGS84).
      schema:
        type: integer
        default: 4326
      example: 4326
...
  schemas:
    GazetteerResponse:
      type: object
      description: Reverse-Geocoding- und Peilungs-Ergebnis für eine Koordinate.
      allOf:
        - type: object
          required:
            - coordinate
          properties:
            coordinate:
              $ref: '#/components/schemas/Coordinate'
            wgs84:
              allOf:
                - $ref: '#/components/schemas/Wgs84Coordinate'
              nullable: true
              description: >-
                Der Abfragepunkt in WGS84 (lon/lat); bei projiziertem `srid`
                reprojiziert (siehe with-gazetteer / Endpoint-Beschreibung).
        - $ref: '#/components/schemas/GazetteerData'
...
    Wgs84Coordinate:
      type: object
      description: >-
        Eine WGS84-Koordinate (EPSG:4326) als lon/lat. Explizit geografisch (nicht
        x/y/srid), damit nachgelagerte Services direkt damit rechnen/speichern können.
      properties:
        lon:
          type: number
          format: double
          description: Längengrad (WGS84).
        lat:
          type: number
          format: double
          description: Breitengrad (WGS84).
      required:
        - lon
        - lat
    Coordinate:
      type: object
      description: Geografische Koordinate
      properties:
        x:
          type: number
          format: double
          description: X-Koordinate (Longitude bei WGS84)
        y:
          type: number
          format: double
          description: Y-Koordinate (Latitude bei WGS84)
        srid:
          type: integer
          description: Spatial Reference ID
      required:
        - x
        - y
        - srid
...
    License:
      type: object
      description: Lizenzinformationen
      properties:
        name:
          type: string
          description: Name der Lizenz
        url:
          type: string
          format: uri
          description: URL zur Lizenz
        attribution:
          type: string
          description: Attributionstext
...
    Source:
      type: object
      description: Datenquellen-Informationen
      properties:
        id:
          type: string
          description: Eindeutige ID der Datenquelle
        name:
          type: string
          description: Anzeigename
        path:
          type: string
          description: Dateipfad
        size:
          type: integer
          format: int64
          description: Dateigröße in Bytes
        layer_count:
          type: integer
          description: Anzahl der Layer
        indexed:
          type: boolean
          description: Alle Indizes erstellt
        ready:
          type: boolean
          description: Bereit für Abfragen
        loaded_at:
          type: string
          format: date-time
          description: Zeitpunkt des Ladens
        last_queried:
          type: string
          format: date-time
          description: Zeitpunkt der letzten Abfrage
        license:
          allOf:
            - $ref: '#/components/schemas/License'
          description: >-
            Lizenz/Attribution der Datenquelle, sofern im Paket hinterlegt.
            Bei GeoPackages ist es die gpkg_metadata-Zeile mit
            mime_type='application/json' UND
            md_standard_uri='https://ortus.dev/schema/dataset-metadata.json'
            (JSON unter anderer URI wird ignoriert); bei Raster-Bundles der
            license-Block des Manifests. Fehlt, wenn das Paket keine Lizenz
            mitführt.
      required:
        - id
        - name
        - path
        - size
        - layer_count
        - indexed
        - ready
    SourceList:
      type: object
      description: Liste von Datenquellen
      properties:
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
        count:
          type: integer
          description: Anzahl der Datenquellen
      required:
        - sources
        - count
...
    Error:
      type: object
      description: Fehlermeldung
      properties:
        error:
          type: string
          description: HTTP-Statustext
        message:
          type: string
          description: Detaillierte Fehlermeldung
      required:
        - error
        - message